refactor(output): group ascii and json outputs
