+++ src/output/json/react.ts
@@ -2,18 +2,18 @@ import {
   getFilteredUsages,
   getReactUsageEntries,
   getReactUsageRoots,
-} from '../analyzers/react/queries.js'
+} from '../../analyzers/react/queries.js'
 import type {
   ReactUsageEntry,
   ReactUsageFilter,
   ReactUsageGraph,
-} from '../types.js'
-import { toDisplayPath } from '../utils/to-display-path.js'
+} from '../../types.js'
+import { toDisplayPath } from '../../utils/to-display-path.js'
 
 interface SerializedReactUsageNode {
   readonly id: string
   readonly name: string
-  readonly symbolKind: import('../types.js').ReactSymbolKind | 'circular'
+  readonly symbolKind: import('../../types.js').ReactSymbolKind | 'circular'
   readonly filePath: string
   readonly exportNames: readonly string[]
   readonly usages: readonly SerializedReactUsageEdge[]
@@ -28,7 +28,7 @@ interface SerializedReactUsageEntry {
 }
 
 interface SerializedReactUsageEdge {
-  readonly kind: import('../types.js').ReactUsageEdge['kind']
+  readonly kind: import('../../types.js').ReactUsageEdge['kind']
   readonly targetId: string
   readonly node: SerializedReactUsageNode
 }
